Introduce a new model: LargeAnswerBucketModel and two new jobs: SplitLargeAnswerBucketsJob and CleanupLargeBucketLabelsFromNewAnswersJob. This new model in conjunction with the first new job go through all answers in the old answers model and split them into shards of 100 answers pre-migration. The answer migration job has been updated to ignore old item buckets if there are corresponding shrad buckets. It will migrate these shard buckets, instead, in parallel. If one fails, the answer migration cleanup job has been updated along with the new answer structure itself to keep track of which bucket was used to migrate that answer, and will cleanup specific answers which were part of a bucket that failed. Whether a bucket failed is tracked using changes to the MigratedAnswerModel. Finally, the second new job is used to remove the large bucket IDs that were added to the new answer structure as part of this cataloging.
Note: This commit includes a variety of new tests to verify everything
is working. One of the tests
(test_reduce_shard_failure_with_large_answers_properly_recovers) has
effectively been verified to pass, but does not currently pass due to
strange behavior where the answer migration and cleanup large bucket
labels jobs do not seem to be properly committing their changes back to
the test. A few hours of debugging did not yield insight as to why this
is, and the logging used to debug is being kept in the codebase as part
of this commit. Any other breakages introduced by this change and
removal of the logs will happen in the immediately followed commit.